@@ -1,6 +1,6 @@
 {
   "name": "check4updates",
-  "version": "1.3.2",
+  "version": "1.3.3-0",
   "description": "Check and update package dependencies.",
   "keywords": [
     "c4u",
@@ -26,54 +26,51 @@
   },
   "license": "MIT",
   "author": "spurreiter",
-  "main": "src",
+  "main": "./src/index.js",
+  "types": "./types/index.d.ts",
   "bin": {
     "c4u": "./bin/c4u.js"
   },
-  "directories": {
-    "test": "test"
-  },
   "files": [
     "bin",
-    "src"
+    "src",
+    "types"
   ],
   "scripts": {
-    "all": "npm-run-all clean pretty lint test",
-    "changelog": "npx conventional-changelog -p eslint -i CHANGELOG.md -s -r1",
-    "clean": "rimraf coverage .nyc_output",
+    "all": "npm-run-all clean pretty lint test types",
+    "ci": "CI=true npm run all",
+    "changelog": "npx conventional-changelog -p eslint -i CHANGELOG.md -s -r1 -u",
+    "clean": "rimraf coverage types",
     "coverage": "c8 -r text npm test",
     "lint": "eslint",
     "pretty": "prettier -w src/*.js -w test/*.js",
-    "test": "mocha"
+    "test": "mocha",
+    "types": "tsc"
   },
   "dependencies": {
     "asyncc-promise": "^1.2.0",
     "chalk": "^4.1.2",
-    "debug": "^4.3.5",
-    "hosted-git-info": "^7.0.2",
-    "ip": "^2.0.1",
+    "debug": "^4.3.7",
+    "hosted-git-info": "^8.0.0",
     "lodash.get": "^4.4.2",
-    "lru-cache": "^11.0.0",
-    "pacote": "^18.0.6",
+    "pacote": "^20.0.0",
     "progress": "^2.0.3",
-    "semver": "^7.6.2",
-    "semver-utils": "^1.1.4"
+    "semver": "^7.6.3"
   },
   "devDependencies": {
-    "@eslint/js": "^9.7.0",
+    "@eslint/js": "^9.15.0",
+    "@types/node": "^22.9.0",
     "c8": "^10.1.2",
-    "eslint": "9.x",
-    "globals": "^15.8.0",
-    "mocha": "^10.6.0",
-    "npm-run-all2": "^6.2.2",
-    "prettier": "^3.3.2",
+    "eslint": "^9.15.0",
+    "globals": "^15.12.0",
+    "mocha": "^10.8.2",
+    "npm-run-all2": "^7.0.1",
+    "prettier": "^3.3.3",
     "rimraf": "^6.0.1",
-    "shelljs": "^0.8.5"
+    "shelljs": "^0.8.5",
+    "typescript": "^5.6.3"
   },
   "c4uIgnore": {
     "chalk": "^4 // version >=5 uses ESM"
-  },
-  "__notes": [
-    "dependencies.ip : ip@2.0.0 moderate vulnerability; remove later"
-  ]
+  }
 }
