@@ -1,5 +1,5 @@
+export type Packages = import("./PckgJson").Packages;
 /** @typedef {import('./PckgJson').Packages} Packages */
-
 /**
  * filter include and excludes
  * @param {object} param1
@@ -10,24 +10,10 @@
  * @param {RegExp} [param1.filterInv] - inverse regular expression to filter for package names
  * @returns {Packages}
  */
-const incexc = ({
-  packages,
-  include = [],
-  exclude = [],
-  filter,
-  filterInv
-}) => {
-  return Object.keys(packages)
-    .filter((p) => (filter ? filter.test(p) : true))
-    .filter((p) => (filterInv ? !filterInv.test(p) : true))
-    .filter((p) => (include.length ? include.includes(p) : true))
-    .filter((p) => (exclude.length ? !exclude.includes(p) : true))
-    .reduce((o, curr) => {
-      if (packages[curr] !== undefined) o[curr] = packages[curr]
-      return o
-    }, {})
-}
-
-module.exports = {
-  incexc
-}
+export function incexc({ packages, include, exclude, filter, filterInv }: {
+    packages: Packages;
+    include?: string[] | undefined;
+    exclude?: string[] | undefined;
+    filter?: RegExp | undefined;
+    filterInv?: RegExp | undefined;
+}): Packages;
