@@ -1,15 +1,21 @@
-const fs = require('fs')
-const { promisify } = require('util')
+const fsp = require('fs/promises')
 const { resolve } = require('path')
 const semver = require('semver')
 
-const fsReadFile = promisify(fs.readFile)
-const fsWriteFile = promisify(fs.writeFile)
+/**
+ * @typedef {{}|Record<string,string>} Packages
+ */
 
 class PckgJson {
+  /**
+   * @param {{
+   *  dirname?: string
+   *  filename?: string
+   * }} param0
+   */
   constructor({ dirname, filename = 'package.json' } = {}) {
     this.dirname = dirname || process.cwd()
-    this.filename = resolve(dirname, filename)
+    this.filename = resolve(this.dirname, filename)
     this.content = undefined
   }
 
@@ -29,6 +35,7 @@ class PckgJson {
    */
   _extract(_content) {
     const packages = {}
+    // @ts-expect-error
     this.fields.forEach((dep) => {
       Object.entries(this.content[dep] || {}).forEach(([pckg, version]) => {
         packages[pckg] = version
@@ -41,6 +48,7 @@ class PckgJson {
    * @private
    */
   _merge(content, packages) {
+    // @ts-expect-error
     this.fields.forEach((dep) => {
       Object.entries(this.content[dep] || {}).forEach(([pckg, _version]) => {
         const nextVersion = packages[pckg]
@@ -83,13 +91,14 @@ class PckgJson {
    *  dev?: boolean
    *  peer?: boolean
    * }} opts
-   * @returns {Promise<Record<string,string>}
+   * @returns {Promise<Packages>}
    */
   read(opts = {}) {
     if (!opts.prod && !opts.dev && !opts.peer) {
       opts = { prod: true, dev: true, peer: true }
     }
-    return fsReadFile(this.filename, 'utf8')
+    return fsp
+      .readFile(this.filename, 'utf8')
       .then((str) => JSON.parse(str))
       .then((content) => {
         this.fields = this._setFields(opts)
@@ -105,7 +114,7 @@ class PckgJson {
   write(packages = {}) {
     this.content = this._merge(this.content, packages)
     const str = JSON.stringify(this.content, null, 2) + '\n'
-    return fsWriteFile(this.filename, str, 'utf8')
+    return fsp.writeFile(this.filename, str, 'utf8')
   }
 }
 
