@@ -47,20 +47,20 @@ const _sortVersions = (a, b) => {
 }
 
 /**
- * @typedef {Object} FoundVersions
- * @property {string} type - version range char `^`, '~'
- * @property {string} max - max available version
- * @property {string} latest - latest available version
- * @property {string} major - matching major version
- * @property {string} minor - matching minor version
- * @property {string} patch - matching patch version
+ * @typedef {object} FoundVersions
+ * @property {string} wildcard version range char `^`, '~'
+ * @property {string} max max available version
+ * @property {string} [latest] latest available version
+ * @property {string} major matching major version
+ * @property {string} minor matching minor version
+ * @property {string} patch matching patch version
  */
 
 /**
  * @param {string[]} versions - list of semver versions
  * @param {string} range - semver range
- * @param {string} latest - latest package version
- * @returns {FoundVersions}
+ * @param {string} [latest] - latest package version
+ * @returns {FoundVersions|null}
  */
 const maxSatisfying = (versions = [], range, latest) => {
   const _range = semver.validRange(range)
